fix: Удалил трейты(trait)как пережиток прошлого стиля программирования
