+++ modules/wm/b24/Query.php
@@ -6,12 +6,13 @@ use Bitrix24\B24Object;
 use wm\b24tools\b24Tools;
 use Yii;
 use yii\base\Component;
+use yii\base\NotSupportedException;
 use yii\helpers\ArrayHelper;
 
 //Код не универсален а направлен на смарт процессы стоит перенести в другой класс
 class Query extends Component implements QueryInterface {
 
-    use QueryTrait;
+
     public $select;
     public $selectOption;
     public $distinct;
@@ -31,8 +32,43 @@ class Query extends Component implements QueryInterface {
     public $queryCacheDuration;
     public $queryCacheDependency;
 
-    //public $emulateExecution;//  изучить тип и место где должен быть.
+    /**
+     * @var string|array|ExpressionInterface|пустое условие запроса. Это относится к предложению WHERE в операторе SQL.
+     * Например, `['возраст' => 31, 'команда' => 1]`.
+     * @see where() для корректного синтаксиса при указании этого значения.
+     */
+    public $where;
+    /**
+     * @var int|ExpressionInterface|null maximum number of records to be returned. May be an instance of [[ExpressionInterface]].
+     * If not set or less than 0, it means no limit.
+     */
+    public $limit;
+    /**
+     * @var int|ExpressionInterface|null zero-based offset from where the records are to be returned.
+     * May be an instance of [[ExpressionInterface]]. If not set or less than 0, it means starting from the beginning.
+     */
+    public $offset;
+    /**
+     * @var array|null how to sort the query results. This is used to construct the ORDER BY clause in a SQL statement.
+     * The array keys are the columns to be sorted by, and the array values are the corresponding sort directions which
+     * can be either [SORT_ASC](https://www.php.net/manual/en/array.constants.php#constant.sort-asc)
+     * or [SORT_DESC](https://www.php.net/manual/en/array.constants.php#constant.sort-desc).
+     * The array may also contain [[ExpressionInterface]] objects. If that is the case, the expressions
+     * will be converted into strings without any change.
+     */
+    public $orderBy;
+    /**
+     * @var string|callable|null the name of the column by which the query results should be indexed by.
+     * This can also be a callable (e.g. anonymous function) that returns the index value based on the given
+     * row data. For more details, see [[indexBy()]]. This property is only used by [[QueryInterface::all()|all()]].
+     */
     public $indexBy;
+    /**
+     * @var bool whether to emulate the actual query execution, returning empty or false results.
+     * @see emulateExecution()
+     * @since 2.0.11
+     */
+    public $emulateExecution = false;
 
 
 
@@ -401,10 +437,11 @@ class Query extends Component implements QueryInterface {
     public function where($condition, $params = [])
     {
         $this->where = $this->conditionPrepare($condition);
-        Yii::warning($this->where, '$this->where');
         $this->addParams($params);
         return $this;
     }
+
+
 
 //    public function from($tables)
 //    {
@@ -687,5 +724,393 @@ class Query extends Component implements QueryInterface {
 //
 //        return $command;
 //    }
+
+
+
+
+
+    /**
+     * Sets the [[indexBy]] property.
+     * @param string|callable $column the name of the column by which the query results should be indexed by.
+     * This can also be a callable (e.g. anonymous function) that returns the index value based on the given
+     * row data. The signature of the callable should be:
+     *
+     * ```php
+     * function ($row)
+     * {
+     *     // return the index value corresponding to $row
+     * }
+     * ```
+     *
+     * @return $this the query object itself
+     */
+    public function indexBy($column)
+    {
+        $this->indexBy = $column;
+        return $this;
+    }
+
+    /**
+     * Adds an additional WHERE condition to the existing one.
+     * The new condition and the existing one will be joined using the 'AND' operator.
+     * @param string|array|ExpressionInterface $condition the new WHERE condition. Please refer to [[where()]]
+     * on how to specify this parameter.
+     * @return $this the query object itself
+     * @see where()
+     * @see orWhere()
+     */
+    public function andWhere($condition)
+    {
+        \Yii::warning($condition, 'andWhere($condition)');
+        $condition = $this->conditionPrepare($condition);
+        if ($this->where === null) {
+            $this->where = $condition;
+        } else {
+            $this->where = array_merge($this->where, $condition);
+        }
+        \Yii::warning($this->where, '$this->where3');
+        return $this;
+    }
+
+    public function conditionPrepare($condition){
+        if(array_key_exists(0, $condition)){
+            if(count($condition)==3){
+                $arr = [];
+                $operator = array_shift($condition);
+                $arr[$operator.$condition[0]] = $condition[1];
+                return $arr;
+            }
+            return [];
+        }else{
+            return $condition;
+        }
+
+    }
+
+//    public function conditionPrepare($condition){
+//        if(count($condition)==3){
+//            $operator = array_shift($condition);
+//            $condition[0] = $operator.$condition[0];
+//        }
+//        return $condition;
+//    }
+
+    /**
+     * Adds an additional WHERE condition to the existing one.
+     * The new condition and the existing one will be joined using the 'OR' operator.
+     * @param string|array|ExpressionInterface $condition the new WHERE condition. Please refer to [[where()]]
+     * on how to specify this parameter.
+     * @return $this the query object itself
+     * @see where()
+     * @see andWhere()
+     */
+    public function orWhere($condition)
+    {
+        if ($this->where === null) {
+            $this->where = $condition;
+        } else {
+            $this->where = ['or', $this->where, $condition];
+        }
+
+        return $this;
+    }
+
+    /**
+     * Sets the WHERE part of the query but ignores [[isEmpty()|empty operands]].
+     *
+     * This method is similar to [[where()]]. The main difference is that this method will
+     * remove [[isEmpty()|empty query operands]]. As a result, this method is best suited
+     * for building query conditions based on filter values entered by users.
+     *
+     * The following code shows the difference between this method and [[where()]]:
+     *
+     * ```php
+     * // WHERE `age`=:age
+     * $query->filterWhere(['name' => null, 'age' => 20]);
+     * // WHERE `age`=:age
+     * $query->where(['age' => 20]);
+     * // WHERE `name` IS NULL AND `age`=:age
+     * $query->where(['name' => null, 'age' => 20]);
+     * ```
+     *
+     * Note that unlike [[where()]], you cannot pass binding parameters to this method.
+     *
+     * @param array $condition the conditions that should be put in the WHERE part.
+     * See [[where()]] on how to specify this parameter.
+     * @return $this the query object itself
+     * @see where()
+     * @see andFilterWhere()
+     * @see orFilterWhere()
+     */
+    public function filterWhere(array $condition)
+    {
+        $condition = $this->filterCondition($condition);
+        if ($condition !== []) {
+            $this->where($condition);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Adds an additional WHERE condition to the existing one but ignores [[isEmpty()|empty operands]].
+     * The new condition and the existing one will be joined using the 'AND' operator.
+     *
+     * This method is similar to [[andWhere()]]. The main difference is that this method will
+     * remove [[isEmpty()|empty query operands]]. As a result, this method is best suited
+     * for building query conditions based on filter values entered by users.
+     *
+     * @param array $condition the new WHERE condition. Please refer to [[where()]]
+     * on how to specify this parameter.
+     * @return $this the query object itself
+     * @see filterWhere()
+     * @see orFilterWhere()
+     */
+    public function andFilterWhere(array $condition)
+    {
+//        \Yii::warning('andFilterWhere');
+        $condition = $this->filterCondition($condition);
+        if ($condition !== []) {
+            $this->andWhere($condition);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Adds an additional WHERE condition to the existing one but ignores [[isEmpty()|empty operands]].
+     * The new condition and the existing one will be joined using the 'OR' operator.
+     *
+     * This method is similar to [[orWhere()]]. The main difference is that this method will
+     * remove [[isEmpty()|empty query operands]]. As a result, this method is best suited
+     * for building query conditions based on filter values entered by users.
+     *
+     * @param array $condition the new WHERE condition. Please refer to [[where()]]
+     * on how to specify this parameter.
+     * @return $this the query object itself
+     * @see filterWhere()
+     * @see andFilterWhere()
+     */
+    public function orFilterWhere(array $condition)
+    {
+        $condition = $this->filterCondition($condition);
+        if ($condition !== []) {
+            $this->orWhere($condition);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Removes [[isEmpty()|empty operands]] from the given query condition.
+     *
+     * @param array $condition the original condition
+     * @return array the condition with [[isEmpty()|empty operands]] removed.
+     * @throws NotSupportedException if the condition operator is not supported
+     */
+    protected function filterCondition($condition)
+    {
+        if (!is_array($condition)) {
+            return $condition;
+        }
+
+        if (!isset($condition[0])) {
+            \Yii::warning($condition, 'if (!isset($condition[0]))');
+            // hash format: 'column1' => 'value1', 'column2' => 'value2', ...
+            foreach ($condition as $name => $value) {
+                if ($this->isEmpty($value)) {
+                    unset($condition[$name]);
+                }
+            }
+
+            return $condition;
+        }
+
+        // operator format: operator, operand 1, operand 2, ...
+
+        $operator = array_shift($condition);
+
+        switch (strtoupper($operator)) {
+            case 'NOT':
+            case 'AND':
+            case 'OR':
+                foreach ($condition as $i => $operand) {
+                    $subCondition = $this->filterCondition($operand);
+                    if ($this->isEmpty($subCondition)) {
+                        unset($condition[$i]);
+                    } else {
+                        $condition[$i] = $subCondition;
+                    }
+                }
+
+                if (empty($condition)) {
+                    return [];
+                }
+                break;
+            case 'BETWEEN':
+            case 'NOT BETWEEN':
+                if (array_key_exists(1, $condition) && array_key_exists(2, $condition)) {
+                    if ($this->isEmpty($condition[1]) || $this->isEmpty($condition[2])) {
+                        return [];
+                    }
+                }
+                break;
+            default:
+//                \Yii::warning($condition, 'default');
+                if (array_key_exists(1, $condition) && $this->isEmpty($condition[1])) {
+//                    \Yii::warning($condition, 'default1');
+                    return [];
+                }
+        }
+
+        array_unshift($condition, $operator);
+
+        return $condition;
+    }
+
+    /**
+     * Returns a value indicating whether the give value is "empty".
+     *
+     * The value is considered "empty", if one of the following conditions is satisfied:
+     *
+     * - it is `null`,
+     * - an empty string (`''`),
+     * - a string containing only whitespace characters,
+     * - or an empty array.
+     *
+     * @param mixed $value
+     * @return bool if the value is empty
+     */
+    protected function isEmpty($value)
+    {
+        return $value === '' || $value === [] || $value === null || is_string($value) && trim($value) === '';
+    }
+
+    /**
+     * Sets the ORDER BY part of the query.
+     * @param string|array|ExpressionInterface $columns the columns (and the directions) to be ordered by.
+     * Columns can be specified in either a string (e.g. `"id ASC, name DESC"`) or an array
+     * (e.g. `['id' => SORT_ASC, 'name' => SORT_DESC]`).
+     *
+     * The method will automatically quote the column names unless a column contains some parenthesis
+     * (which means the column contains a DB expression).
+     *
+     * Note that if your order-by is an expression containing commas, you should always use an array
+     * to represent the order-by information. Otherwise, the method will not be able to correctly determine
+     * the order-by columns.
+     *
+     * Since version 2.0.7, an [[ExpressionInterface]] object can be passed to specify the ORDER BY part explicitly in plain SQL.
+     * @return $this the query object itself
+     * @see addOrderBy()
+     */
+    public function orderBy($columns)
+    {
+        $this->orderBy = $this->normalizeOrderBy($columns);
+        return $this;
+    }
+
+    /**
+     * Adds additional ORDER BY columns to the query.
+     * @param string|array|ExpressionInterface $columns the columns (and the directions) to be ordered by.
+     * Columns can be specified in either a string (e.g. "id ASC, name DESC") or an array
+     * (e.g. `['id' => SORT_ASC, 'name' => SORT_DESC]`).
+     *
+     * The method will automatically quote the column names unless a column contains some parenthesis
+     * (which means the column contains a DB expression).
+     *
+     * Note that if your order-by is an expression containing commas, you should always use an array
+     * to represent the order-by information. Otherwise, the method will not be able to correctly determine
+     * the order-by columns.
+     *
+     * Since version 2.0.7, an [[ExpressionInterface]] object can be passed to specify the ORDER BY part explicitly in plain SQL.
+     * @return $this the query object itself
+     * @see orderBy()
+     */
+    public function addOrderBy($columns)//['id'=>4]
+    {
+        $columns = $this->normalizeOrderBy($columns);
+        foreach ($columns as $key=>$value){
+            $temp = [];
+            if($value == SORT_ASC){
+                $temp[$key] = 'ASC';
+            }elseif($value == SORT_DESC){
+                $temp[$key] = 'DESC';
+            }
+            $columns = array_merge($columns, $temp);
+        }
+        if ($this->orderBy === null) {
+            $this->orderBy = $columns;
+        } else {
+            $this->orderBy = array_merge($this->orderBy, $columns);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Normalizes format of ORDER BY data.
+     *
+     * @param array|string|ExpressionInterface $columns the columns value to normalize. See [[orderBy]] and [[addOrderBy]].
+     * @return array
+     */
+    protected function normalizeOrderBy($columns)
+    {
+        if ($columns instanceof ExpressionInterface) {
+            return [$columns];
+        } elseif (is_array($columns)) {
+            return $columns;
+        }
+
+        $columns = preg_split('/\s*,\s*/', trim($columns), -1, PREG_SPLIT_NO_EMPTY);
+        $result = [];
+        foreach ($columns as $column) {
+            if (preg_match('/^(.*?)\s+(asc|desc)$/i', $column, $matches)) {
+                $result[$matches[1]] = strcasecmp($matches[2], 'desc') ? SORT_ASC : SORT_DESC;
+            } else {
+                $result[$column] = SORT_ASC;
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * Sets the LIMIT part of the query.
+     * @param int|ExpressionInterface|null $limit the limit. Use null or negative value to disable limit.
+     * @return $this the query object itself
+     */
+    public function limit($limit)
+    {
+        $this->limit = $limit;
+        return $this;
+    }
+
+    /**
+     * Устанавливает часть запроса OFFSET.
+     * @param int|ExpressionInterface|null $offset смещение. Используйте нулевое или отрицательное значение, чтобы отключить смещение.
+     * @return $это сам объект запроса
+     */
+    public function offset($offset)
+    {
+        //TODO Переписать-проверить
+        $this->offset = $offset;
+        return $this;
+    }
+
+    /**
+     * Sets whether to emulate query execution, preventing any interaction with data storage.
+     * After this mode is enabled, methods, returning query results like [[QueryInterface::one()]],
+     * [[QueryInterface::all()]], [[QueryInterface::exists()]] and so on, will return empty or false values.
+     * You should use this method in case your program logic indicates query should not return any results, like
+     * in case you set false where condition like `0=1`.
+     * @param bool $value whether to prevent query execution.
+     * @return $this the query object itself.
+     * @since 2.0.11
+     */
+    public function emulateExecution($value = true)
+    {
+        $this->emulateExecution = $value;
+        return $this;
+    }
 
 }
